Rename tranquillo -> tranquilo
