+++ environment.yml
@@ -30,6 +30,7 @@ dependencies:
   - pytest-cov
   - pytest-xdist
   - scipy>=1.2.1
+  - scikit-learn
   - seaborn
   - sphinx=2.4.4
   - pydata-sphinx-theme>=0.3.0
